ci: run sonarcloud via pull_request_target
`on.pull_request` executes the workflow out of a forked repository. This means the `SONAR_TOKEN` secret will not be available and therefore the workflow will fail.
By changing it to `on.pull_request_target`, the workflow definition (and code) from the trusted original repository is run and the secret will be available.

see https://github.blog/2020-08-03-github-actions-improvements-for-fork-and-pull-request-workflows/