feat(vlan): Add vlan tag support for VM reconcile. (#133)
* feat(vlan): Add vlan tag support for VM reconcile.

This commit implement VLAN tag support when creating - or more in general reconciling - a ProxmoxMachine. This feature implements a new optional value in the ProxmoxMachine CRS, defined in the NetworkDevice type. The value is an integer, delimited in the range 1-4094.

Tests have been also implemented to prevent regression, and e2e tests have been conducted creating both cluster using the new property and without - to check past compat.

* test(vlan): Add more testing and fix validation

* fix(test-vlan): Fix regression in test caused by a typo

* misc: Add more entry to gitignore

* refactor: Removed unused validator and refactor variable name

* refactor: Refactor variable name