bootstrap: default to `format=cloud-config`
Not every bootstrap provider sets the `format` key and i wasn't able to find any requirement for it in the capi contract.

With this change we default to cloud-config, as it was without this check, to restore functionality with other non-kubeadm bootstrap providers.

Takes care of cases where the CAPMOX controller errored out with
```
err="failed to reconcile VM: error retrieving format data: secret `format` key is missing"
```